+++ apps/client/src/app/pages/home/components/sections/AboutSection.tsx
