+++ apps/client/src/app/pages/home/components/sections/FunctionsSection.tsx
