+++ apps/client/tsconfig.json
@@ -63,14 +63,8 @@
       "@home-assets/*": [
         "src/app/pages/home/assets/*"
       ],
-      "@home-sections/*": [
-        "src/app/pages/home/sections/*"
-      ],
       "@home-components/*":[
-        "src/app/pages/home/header/components/*"
-      ],
-      "@header-components/*":[
-        "src/app/pages/home/components/header/components/*"
+        "src/app/pages/home/components/*"
       ],
     }
   },
